+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Added
+- Post modified line exec command ([#888](https://github.com/MitMaro/git-interactive-rebase-tool/pull/888))
+
+
 ## [2.3.0] - 2023-07-19
 ### Added
 - Support for update-ref action ([#801](https://github.com/MitMaro/git-interactive-rebase-tool/pull/801))